requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
psycopg[binary]>=3.2
django-environ>=0.11
pytest>=8.0
pytest-django>=4.8
factory-boy>=3.3
Faker>=25.0
black>=24.0
ruff>=0.5
celery>=5.4
django-celery-beat>=2.6
redis>=5.0
flower>=2.0